feat: :sparkles: Add styling to button, search-bar, and animations
Added features for search-bar and button like focus-within, hover. Added features for button like color changes for hover, active, visited, and animation to the svg in the button icon when hovered. Added Misty Card feature. Slight mist revolving around the exterior of the card. Subtle but gives the card life. New class -> .cardS added for mist features. Class .stage added for background animated gradient for button.

When the cover was hovered previously, there was a glitch feel. The animations of background color from rgba(... 0%) to rgba(... 100%) was the reason. Removed animation transition for opacity from 1 to 0 for the mist feature, and background for .classS