perf: :rotating_light: Fixes use of `==` and `!=`. Now using `===` and `!==` for JavaScript
Good practice to use type-safety equality operators `===` and `!==` instead of  the regular `==` and `!=`. For example, it could be a problem if the regular occurs in`a == b` as it's hard to spot. > "The most notable difference between this operator and the equality (==) operator is that if the operands are of different types, the == operator attempts to convert them to the same type before comparing." Reference - "JS-0050" — Deepsource